Update AQI Generic Messages
- Update AQI Generic Messages, category meaning.
